# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Travelode API
host: 127.0.0.1:28825
basePath: /
schemes:
  - http
  - https
consumes:
  - application/json
produces:
  - application/json
...
paths:
...
  /media:
    x-swagger-router-controller: media

    get:
      tags:
        - Media
      summary: Returns media list
      operationId: findMedias
      parameters:
        - name: page
          in: query
          description: The page number to get. Default to 0.
          required: false
          type: string
        - name: size
          in: query
          description: The size of the page to get. Default to 50.
          required: false
          type: string
        - name: type
          in: query
          description: Find media by type (0=all, 1=photo, 2=video)
          required: true
          type: number
          enum:
            - 0
            - 1
            - 2
          default: 0
        - name: userId
          in: query
          description: Find by owner
          required: false
          type: number
        - name: locationId
          in: query
          description: Find by location
          required: false
          type: number
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/MediaListResponse"
        # responses may fall through to errors
        "500":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

    post:
      tags:
        - Media
      summary: Uploads a media
      operationId: createMedia
      consumes:
        - multipart/form-data
        - application/x-www-form-urlencoded
      produces:
          - text/plain
      parameters:
        - in: formData
          name: userId
          required: true
          type: number
          description: The media owner
        - in: formData
          name: tripMedia
          required: true
          type: file
          description: The media to upload
      responses:
        "201":
          description: Created
          schema:
            $ref: "#/definitions/CreatedResponse"
        "404":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
        # responses may fall through to errors
        "500":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /media/{id}:
    x-swagger-router-controller: media

    get:
      tags:
        - Media
      summary: Returns a single media
      operationId: getMediaById
      parameters:
        - name: id
          in: path
          description: The Media Id
          required: true
          type: integer
          minimum: 1
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/MediaListResponse"
        "404":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
        "500":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

    put:
      tags:
        - Media
      summary: Update a media
      operationId: updateMediaById
      parameters:
        - name: id
          in: path
          description: The media id to update
          required: true
          type: integer
          minimum: 1
        - name: mediaData
          in: body
          description: The media data to update
          required: true
          schema:
            type: object
            properties:
              userId:
                type: number
              locationId:
                type: number
              sizeX:
                type: number
              sizeY:
                type: number
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/SuccessResponse"
        "400":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
        "404":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
        "500":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

    delete:
      tags:
        - Media
      summary: Delete a media
      operationId: deleteMediaById
      parameters:
        - name: id
          in: path
          description: The media id to delete
          required: true
          type: integer
          minimum: 1
      responses:
        "204":
          description: Media is removed
        "404":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
        "500":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
  SuccessResponse:
    required:
      - message
    properties:
      message:
        type: string
  CreatedResponse:
    required:
      - message
      - id
    properties:
      message:
        type: string
      id:
        type: number
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
...
  Media:
    type: object
    properties:
      id:
        type: number
        description: Media Id
      type:
        type: string
        description: Type of the media (photo or video)
      userId:
        type: number
        description: Media owner userId
      uploaded:
        description: Media uploading date time
        type: string
        format: date-time
      updated:
        description: Media update date time
        type: string
        format: date-time
      created:
        description: Media creation date time
        type: string
        format: date-time
      sizeX:
        type: number
        description: Media size width in px
      sizeY:
        type: number
        description: Media size height in px
      url:
        type: string
        description: Url for the media
  MediaListResponse:
    required:
      - medias
    properties:
      # The array of medias
      medias:
        type: array
        items:
          $ref: "#/definitions/Media"
...